fix(UserSegmentObject): parse large int values

+++ src/main/java/lol/pbu/zendesk/client/UserSegmentsClient.java
@@ -15,6 +15,7 @@ import reactor.core.publisher.Mono;
 import static io.micronaut.http.HttpHeaders.ACCEPT;
 
 /**
+ * See <a href="https://developer.zendesk.com/api-reference/help_center/help-center-api/user_segments/">User Segments</a>
  * @author Jonathan Zollinger
  * @since 0.0.1
  */
@@ -46,14 +47,10 @@ public interface UserSegmentsClient {
      * @return Response when the object was deleted (status code 204)
      */
     @Delete("/api/v2/help_center/user_segments/{user_segment_id}")
-    Mono<HttpResponse<Void>> deleteUserSegment(
-            @PathVariable("user_segment_id") @NotNull Integer userSegmentId
-    );
+    Mono<HttpResponse<Void>> deleteUserSegment(@PathVariable("user_segment_id") @NotNull Long userSegmentId);
 
     /**
-     * List Sections with User Segment<br>
-     * <p>This endpoint supports pagination as described in <a href="https://developer.zendesk.com//api-reference/help_center/help-center-api/help-center-api/#pagination">
-     *     Pagination</a>.</p>
+     * List Sections for a given user segment<br>
      * <h4 id="allowed-for">Allowed for</h4>
      * <ul>
      *   <li>Help Center managers</li>
@@ -63,31 +60,21 @@ public interface UserSegmentsClient {
      * @return OK Response (status code 200)
      */
     @Get("/api/v2/help_center/user_segments/{user_segment_id}/sections")
-    Mono<@Valid SectionsResponse> listUserSegmentSections(
-            @PathVariable("user_segment_id") @NotNull Integer userSegmentId
-    );
+    Mono<@Valid SectionsResponse> listUserSegmentSections(@PathVariable("user_segment_id") @NotNull Long userSegmentId);
 
     /**
      * List Topics with User Segment<br>
-     * Lists the topics that use the specified user segment.  This endpoint supports pagination as described in
-     * <a href="https://developer.zendesk.com/api-reference/help_center/help-center-api/help-center-api/#pagination">Pagination</a>.
+     * Lists the topics that use the specified user segment.
      * <h4>Allowed for</h4>
      * <ul>
      * <li>Help Center managers  </li></ul>
-     * <h4>Pagination</h4>
-     * Cursor pagination (recommended)
-     * <ul>
-     * <li>Offset pagination
-     * <br>See <a href="https://developer.zendesk.com/api-reference/introduction/pagination/">Pagination</a>.</li></li>
-     * </ul>
+     *
      *
      * @param userSegmentId The unique ID of the user segment
      * @return OK Response (status code 200)
      */
     @Get("/api/v2/help_center/user_segments/{user_segment_id}/topics")
-    Mono<@Valid TopicsResponse> listUserSegmentTopics(
-            @PathVariable("user_segment_id") @NotNull Integer userSegmentId
-    );
+    Mono<@Valid TopicsResponse> listUserSegmentTopics(@PathVariable("user_segment_id") @NotNull Long userSegmentId);
 
     /**
      * List User Segments<br>
@@ -98,44 +85,30 @@ public interface UserSegmentsClient {
      * The {@code /api/v2/help_center/users/{user_id}/user_segments.json} endpoint returns the list of user segments
      * that a particular user belongs to. This is the only list endpoint that agents have access to. When an agent
      * makes a request to this endpoint with another user's id, the response only includes user segments that the
-     * requesting agent also belongs to.  These endpoints support pagination, as described in the
-     * <a href="https://developer.zendesk.com/api-reference/introduction/pagination/">pagination documentation</a>.
+     * requesting agent also belongs to.
      * <h4>Allowed for</h4>
      * <ul>
-     * <li>Help Center managers s  Agents  </li></ul>
-     * <h4>Pagination  </h4>:
-     * <ul>
-     * <li>Cursor pagination (recommended)</li>
-     * <li>Offset pagination
-     * <br>See <a href="https://developer.zendesk.com/api-reference/introduction/pagination/">Pagination</a>.</li></li>
-     * </ul>
+     * <li>Help Center managers</li>
+     * <li>Agents</li></ul>
+     *
      *
      * @param builtIn Only built_in user segments if true, only custom user segments if false 
      * @return OK response (status code 200)
      */
     @Get("/api/v2/help_center/user_segments")
-    Mono<@Valid UserSegmentsResponse> listUserSegments(
-            @QueryValue("built_in") @Nullable Boolean builtIn
-    );
+    Mono<@Valid UserSegmentsResponse> listUserSegments(@QueryValue("built_in") @Nullable Boolean builtIn);
 
     /**
      * Show User Segment<br>
      * <h4>Allowed for</h4>
      * <ul>
-     * <li>Help Center managers  <h4>Pagination    Cursor pagination (recommended) </h4>:
-     * <ul>
-     * <li>* Offset pagination<br>See <a href="https://developer.zendesk.com/api-reference/introduction/pagination/">Pagination</a>.</li></li>
-     * </ul>
-     *
-     * </ul>
+     * <li>Help Center managers  </li></ul>
      *
      * @param userSegmentId The unique ID of the user segment
      * @return OK Response (status code 200)
      */
     @Get("/api/v2/help_center/user_segments/{user_segment_id}")
-    Mono<@Valid UserSegmentResponse> showUserSegment(
-            @PathVariable("user_segment_id") @NotNull Integer userSegmentId
-    );
+    Mono<@Valid UserSegmentResponse> showUserSegment(@PathVariable("user_segment_id") @NotNull Long userSegmentId);
 
     /**
      * Update User Segment<br>
@@ -149,7 +122,5 @@ public interface UserSegmentsClient {
      * or Bad request Response (status code 400)
      */
     @Put("/api/v2/help_center/user_segments/{user_segment_id}")
-    Mono<@Valid UserSegmentResponse> updateUserSegment(
-            @PathVariable("user_segment_id") @NotNull Integer userSegmentId
-    );
+    Mono<@Valid UserSegmentResponse> updateUserSegment(@PathVariable("user_segment_id") @NotNull Long userSegmentId);
 }
